Checkpoint before assistant change: Create simple tools to perform calculations and print patterns on screen
Adds calculator.c to perform integer addition and mario.c/meow.c for printing patterns.

+++ 1/calculator.c
@@ -6,10 +6,10 @@
 int main(void)
 {
     //prompt user for x
-    float x = get_int("x: ");
+    int x = get_int("x: ");
 
     //prompt user for y
-    float y = get_int("y: ");
+    int y = get_int("y: ");
 
     //add numbers
     int z = x + y;
